完成消息列表接口对接

+++ src/views/user/InBox.vue
@@ -15,11 +15,22 @@
                         type="selection"
                         width="55">
                 </el-table-column>
+                <el-table-column width="1">
+                    <template slot-scope="scope">
+                        <span style="display: none;">{{scope.row.index = scope.$index}}</span>
+                    </template>
+                </el-table-column>
                 <el-table-column
                         label="发件人"
                         width="150">
                     <template slot-scope="scope">
-                        <span class="inbox-message-form"> {{scope.row.from}}</span>
+                        <div class="inbox-message-form">
+                            <el-badge is-dot :hidden="scope.row.isRead">
+                                <template v-if="scope.row.fromId == 'admin'">管理员</template>
+                                <template v-else-if="scope.row.fromId == 'system'">系统消息</template>
+                                <template v-else>用户消息</template>
+                            </el-badge>
+                        </div>
                     </template>
                 </el-table-column>
                 <el-table-column
@@ -39,7 +50,7 @@
                         label="时间"
                         width="180">
                     <template slot-scope="scope">
-                        <span class="inbox-message-date"> {{scope.row.date}}</span>
+                        <span class="inbox-message-date"> {{scope.row.createTime}}</span>
                     </template>
                 </el-table-column>
             </el-table>
@@ -50,7 +61,7 @@
 </template>
 
 <script>
-    import {Table, TableColumn, Button} from "element-ui"
+    import {Table, TableColumn, Button, Badge} from "element-ui"
     import UserMessageInfo from "../../components/user/inbox/UserMessageInfo";
 
     export default {
@@ -59,34 +70,40 @@
             UserMessageInfo,
             [Table.name]: Table,
             [TableColumn.name]: TableColumn,
-            [Button.name]: Button
+            [Button.name]: Button,
+            [Badge.name]: Badge
         },
         data() {
             return {
-                messages: [{
-                    from: "管理员",
-                    title: "实名认证审核进度",
-                    content: "恭喜你，已通过实名认证1",
-                    date: "2020-02-19",
-                    isRead: true,
-                }, {
-                    from: "管理员",
-                    title: "实名认证审核进度",
-                    content: "恭喜你，已通过实名认证2",
-                    date: "2020-02-20",
-                    isRead: false,
-                }],
+                messages: [],
                 showMessage: false,
                 currMessage: {},
                 selectedMessages: [],//已经选择的消息列表
             }
         },
+        created() {
+            this.initMessages();
+        },
         methods: {
-            lookMessage(row) {
+            //初始化消息列表
+            initMessages(){
+                this.$API.getMessages().then( res => {
+                    if(res.data.error == 0){
+                        this.messages = res.data.messages;
+                    }else{
+                        this.$message.error(res.data.error_message);
+                    }
+                })
+            },
+            lookMessage(row) {//查看消息
                 this.showMessage = true;
                 this.currMessage = row;
+                if(!row.isRead){//如果是未读的，就设为已读
+                    this.messages[row.index].isRead = true;
+                    this.$API.readMessage(row.messageId);
+                }
             },
-            closeMessage() {
+            closeMessage() {//关闭消息窗口
                 this.showMessage = false;
                 this.currMessage = {};
             },
@@ -100,17 +117,34 @@
                         title: "是否要删除消息",
                         showCancelButton: true
                     }).then(() => {
-                        let originMessages = JSON.stringify(this.messages);
-                        for (let i = 0; i < this.selectedMessages.length; i++) {
-                            let deleteItem = JSON.stringify(this.selectedMessages[i]);
-                            if (originMessages.indexOf(deleteItem + ",") != -1) {
-                                originMessages = originMessages.replace(deleteItem + ",", "");
-                                continue;
+                        this.$API.deleteMessages(this.selectedMessages).then(res => {
+                            window.console.log(res.data)
+                            if(res.data.error == 0){
+                                //删除成功
+                                this.$message.success("已删除");
+                                let originMessages = JSON.stringify(this.messages);
+                                for (let i = 0; i < this.selectedMessages.length; i++) {
+                                    let deleteItem = JSON.stringify(this.selectedMessages[i]);
+                                    if (originMessages.indexOf(deleteItem + ",") != -1) {
+                                        originMessages = originMessages.replace(deleteItem + ",", "");
+                                        continue;
+                                    }
+                                    originMessages = originMessages.replace(deleteItem, "");
+                                }
+                                //如果有删除最后一个，则有可能会多出一个逗号，替换掉
+                                originMessages = originMessages.replace("},]", "}]");
+                                this.messages = JSON.parse(originMessages);
+                                this.selectedMessages = [];//清空选择的消息列表
+                            }else{
+                                //删除失败,重新获取消息列表
+                                this.$message.error(res.data.error_message);
+                                this.initMessages();
                             }
-                            originMessages = originMessages.replace(deleteItem, "");
-                        }
-                        this.messages = JSON.parse(originMessages);
+                        });
                     }).catch(() => {
+                        //删除失败,重新获取消息列表
+                        this.$message.error("发生意外错误，删除失败");
+                        this.initMessages();
                     })
                 } else {
                     this.$message.info("请先勾选要删除的消息")
@@ -127,6 +161,7 @@
 
     .inbox-message-form {
         font-weight: bold;
+        padding-top: 4px;
     }
 
     .inbox-message-title {
